+++ example/basic-string/app.js
@@ -9,7 +9,7 @@ stage.viewbox(300, 200);
 
 let box = Stage.sprite("box").box().stretch().padding(10).pin("align", 0.5).appendTo(stage);
 
-let number = Stage.string("digit").value("0123456789").pin("align", 0.5).appendTo(box);
+let number = Stage.monotype("digit").value("0123456789").pin("align", 0.5).appendTo(box);
 
 stage.on("click", function (point) {
   let range = math.pow(10, math.random(0, 10) | 0);